Add Clock functionality
tg::Clock::now() can be called to get the current time since that function was called the first time. Ideally, call this function at the beginning of main().
ClockTime is a time type holding a number of seconds and microseconds. It can be used both as a clock time type and as a duration type interchangeably.